requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.1

# Supabase Storage (lightweight, no auth dependencies)
storage3==0.7.7

# File validation
python-magic==0.4.27

# Document processing
python-docx==1.1.0

# Environment management
python-dotenv==1.0.0

# HTTP client for testing
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.0
